quick-commerce-deal-manual/requirements.txt
langchain-community
langchain-openai
openai
python-dotenv
pandas
numpy
faker
plotly
streamlit
sqlalchemy
langchain
langchain-core
pprint